Added: Ability to search Person in a City or State across the multiple AddressBook
